+++ sealed_example/bin/sealed/nullsafe/data/generic/result_no_upper_1.sealed.dart
@@ -18,11 +18,11 @@ part of 'result_no_upper_1.dart';
 abstract class Result<D extends Object?, E extends Object?> {
   const Result._internal();
 
-  factory Result.success({
+  const factory Result.success({
     required D data,
   }) = ResultSuccess<D, E>;
 
-  factory Result.error({
+  const factory Result.error({
     required E exception,
   }) = ResultError<D, E>;