+++ sealed_example/test/sealed/legacy_all.dart
@@ -1,8 +1,0 @@
-// @dart=2.9
-import 'package:test/test.dart';
-
-import 'legacy/all.dart' as all_legacy;
-
-void main() {
-  group('legacy', all_legacy.main);
-}