+++ sealed_example/test/sealed/nullsafe/identity/simple/all.dart
@@ -1,7 +1,0 @@
-import 'package:test/test.dart';
-
-import 'test_weather.dart' as test_weather;
-
-void main() {
-  group('test_weather', test_weather.main);
-}