+++ super_enum_sealed_example/test/legacy_test.dart
@@ -1,6 +1,0 @@
-// @dart=2.9
-import 'legacy/all_legacy.dart' as legacy;
-
-void main() {
-  legacy.main();
-}