+++ super_enum_sealed_example/test/nullsafe_test.dart
@@ -1,5 +1,0 @@
-import 'nullsafe/all_nullsafe.dart' as nullsafe;
-
-void main() {
-  nullsafe.main();
-}